+++ RELEASE_NOTES.md
@@ -5,7 +5,10 @@ Release Notes
 * Container Service (AKS): Support basic SKU for the cluster's load balancer (default is standard).
 * Container Service (AKS): Support for private Kubernetes API access.
 * Container Service (AKS): Support for restricting IP ranges for Kubernetes API access.
+* Functions: Support publishing as a docker container
 * Service bus: Add support for authorization rules.
+* Virtual Machine: Added disable password authentication to Virtual Machine linux configuration
+* Virtual Machine: Added sshKeys and paths to Virtual Machine linux configuration
 
 ## 1.6.2
 * Functions: Support Elastic Premium SKUs for Functions service plans.